Created Checkpoint script as separate script in system scripts
updated checkpoint prefab to not have mesh renderer. created checkpoint to update player prefs when player move through it. renamed door patrol switch to checkpoint.
