+++ Assets/Scripts/System_Scripts/Checkpoint.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b64df5724df98b244b99e86f1fedd78b
+guid: 73d01d0fe01a0fc41adcf6754b1dc0e1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
